+++ src/components/Search/Search.tsx
@@ -18,6 +18,7 @@ import {
   useModal,
   TemplateFooter,
   Footer,
+  useNYPLBreakpoints,
 } from "@nypl/design-system-react-components";
 import { useRouter } from "next/router";
 import { FacetItem, Query } from "~/src/types/DataModel";
@@ -52,6 +53,8 @@ const SearchResults: React.FC<{
   const { isFlagActive } = useFeatureFlags();
 
   const { onClose, onOpen, Modal } = useModal();
+
+  const { isLargerThanMedium } = useNYPLBreakpoints();
 
   const router = useRouter();
 
@@ -232,20 +235,18 @@ const SearchResults: React.FC<{
           </Flex>
         </TemplateContentTop>
         <TemplateContentSidebar>
-          <Button
-            id="filter-button"
-            onClick={onOpen}
-            buttonType="secondary"
-            __css={{
-              width: "100%",
-              display: {
-                base: "block",
-                md: "none",
-              },
-            }}
-          >
-            {`Filters (${filterCount})`}
-          </Button>
+          {!isLargerThanMedium && (
+            <Button
+              id="filter-button"
+              onClick={onOpen}
+              buttonType="primary"
+              sx={{
+                width: "100%",
+              }}
+            >
+              {`Filters (${filterCount})`}
+            </Button>
+          )}
           <Modal
             bodyContent={
               <>
@@ -289,6 +290,22 @@ const SearchResults: React.FC<{
               </>
             }
           />
+          {searchQuery.filters.length > 0 && !isLargerThanMedium && (
+            <Button
+              id="clear-filters-button"
+              buttonType="secondary"
+              type="reset"
+              onClick={() => {
+                changeFilters([]);
+              }}
+              sx={{
+                marginTop: "var(--nypl-space-s)",
+                width: "100%",
+              }}
+            >
+              Clear Filters
+            </Button>
+          )}
           <Form
             id="search-filter-form"
             bg="ui.gray.x-light-cool"